requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
django-storages==1.14.2
boto3==1.34.34
python-dotenv==1.0.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
Pillow==10.2.0
django-environ==0.11.2
django-extensions==3.2.3